♻️파일 분리
